Added working lose condition and pirateship
I added a pirateShip that slowly approaches, just so that I could check so that the lose conditions are actually triggered when the pirateship is close enough to board.

The pirateship needs to be developed further though, seeing as it just comes straight towards you, and can't take damage, or anything like that.