Added icons to cargo crates
Each side of the cargo crates now has a Sprite Renderer and the cargo holds two sprites, one of a cat, showing unknown, and one of the content for when that is revealed.
I also set up a function that can be called on cargo to update the sprite based on the bool "IsRevealed".