Event destroys cargo & cannonBalls
I added an event sent from the gameManager that the projectiles and the cargo subscribes to. And then I built a function that handled destruction for them.
There is still a bug that complains when some cannonballs seems to be looked for even after they are destroyed, and I'll look at that next.