Bugfix CrowbarStation blocking raycast
I think the issue with the crowbar-station was that it's collider blocked the cursor, so you couldn't pick up cargo if it was placed in the very center of it. So I set the crowbar-station to ignoring RayCasts and the problem is gone!

I also changed the code so that the dropspotlight is enabled and disabled, instead of activating the whole gameObject.

So there might be some problem with that later on, if I missed to switch it at some place.